requirements.txt
streamlit
pillow
opencv-python
sqlalchemy
pandas
numpy

